fix mouvement when changing scene
Mouvements controls were not working after moving from a scene to an
other one. Problem's source seems to come from... the texture of the
ground, when adding a new texture (different from the one in the
initial scene) like grass for exemple it make a weird bug with the
controls : the controlers event was reacting and editing XROrigin
position but the player wasn't moving. It's fixed when settings a more
"classiq" texture like a color.