add hand interaction to practices
ISSUE: When moving from catch introduction to catch pratice, balls
aren't at the right place, they stay on the ground not at performance
position. When I remove lines 250-255 (useEffect where Handstate is
created) the bugs seems corrected... The weird thing is that the bug
does not occur in TossPractice which have exactly the same structure.